Add Holiday week. Correct bugs with both test and code

+++ backend/Core/DomainModels/Organization.cs
@@ -9,7 +9,7 @@ public class Organization
     public required string Id { get; set; }
 
     public required string Name { get; set; }
-    public required float HoursPerWorkday { get; set; }
+    public required double HoursPerWorkday { get; set; }
 
     [JsonIgnore] public List<Department> Departments { get; set; }
